+++ src/FleetManager.Api/Controllers/RentalPlanController.cs
@@ -39,7 +39,7 @@ namespace FleetManager.Api.Controllers
         [ProducesResponseType(typeof(ResponseRentalPlanJson), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(ResponseErrorJson), StatusCodes.Status400BadRequest)]
         [ProducesResponseType(typeof(ResponseErrorJson), StatusCodes.Status404NotFound)]
-        public async Task<IActionResult> GetById([FromServices] IGetByIdRentalPlanUseCase useCase, [FromRoute] int id)
+        public async Task<IActionResult> GetById([FromServices] IGetByIdRentalPlanUseCase useCase, [FromRoute] long id)
         {
             var response = await useCase.Execute(id);
 
@@ -49,7 +49,7 @@ namespace FleetManager.Api.Controllers
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(typeof(ResponseErrorJson), StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(ResponseErrorJson), StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> Update([FromServices] IUpdateRentalPlanUseCase useCase, [FromRoute] int id, [FromBody] RequestRentalPlansJson request)
+        public async Task<IActionResult> Update([FromServices] IUpdateRentalPlanUseCase useCase, [FromRoute] long id, [FromBody] RequestRentalPlansJson request)
         {
             await useCase.Execute(id, request);
             return NoContent();
