+++ src/FleetManager.Application/UseCase/ToCompany/Delete/IDeleteCompanyUseCase.cs
@@ -4,5 +4,5 @@ namespace FleetManager.Application.UseCase.ToCompany.Delete;
 
 public interface IDeleteCompanyUseCase
 {
-    Task Execute(int id);
+    Task Execute(long id);
 }
