+++ src/FleetManager.Application/UseCase/ToCompany/GetById/IGetByIdCompanyUseCase.cs
@@ -5,5 +5,5 @@ namespace FleetManager.Application.UseCase.ToCompany.GetById;
 
 public interface IGetByIdCompanyUseCase
 {
-    Task<ResponseCompanyJson> Execute(int id);
+    Task<ResponseCompanyJson> Execute(long id);
 }
