+++ src/FleetManager.Domain/Repositories/ToCompany/ICompanyWriteOnlyRepository.cs
@@ -5,5 +5,5 @@ namespace FleetManager.Domain.Repositories.ToCompany;
 public interface ICompanyWriteOnlyRepository
 {
     Task Add(Company company);
-    Task Delete(int id);
+    Task Delete(long id);
 }
