+++ src/FleetManager.Domain/Repositories/ToRentalPlans/IRentalPlansWriteOnlyRepository.cs
@@ -5,6 +5,6 @@ namespace FleetManager.Domain.Repositories.ToRentalPlans
     public interface IRentalPlansWriteOnlyRepository
     {
         Task Add(RentalPlan rentalPlan);
-        Task Delete(int id);
+        Task Delete(long id);
     }
 }
