+++ src/FleetManager.communication/Requests/RequestRentJson.cs
@@ -3,10 +3,10 @@
     public class RequestRentJson
     {
       
-        public int CompanyId { get; set; }
+        public long CompanyId { get; set; }
         public long ClientId { get; set; }
         public long VehicleId { get; set; }
-        public int RentalPlanId {get ; set ;}
+        public long RentalPlanId {get ; set ;}
         public decimal IncludedKm { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? EndDate { get; set; }
