+++ tests/WebApi.Tests/Resource/CompanyIdentityManager.cs
@@ -5,6 +5,6 @@ namespace WebApi.Tests.Resource
     public class CompanyIdentityManager(Company company)
     {
         private readonly Company _company = company;
-        public int GetById() => _company.Id;
+        public long GetById() => _company.Id;
     }
 }
