+++ tests/WebApi.Tests/ToCompany/Get/GetByIdCompanyTest.cs
@@ -8,7 +8,7 @@ namespace WebApi.Tests.ToCompany.Get
         private const string MEHTOD = "api/Company";
         private readonly HttpClient _;
         private readonly string _userAdmin;
-        private readonly int _companyId;
+        private readonly long _companyId;
         public GetByIdCompanyTest(CustomWebApplicationFactory customWebApplication) : base(customWebApplication)
         {
             _ = customWebApplication.CreateClient();
